requirements.txt
asgiref==3.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.4
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
google-auth==2.40.3
idna==3.10
inflection==0.5.1
mysqlclient==2.2.7
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
rest-framework-simplejwt==0.0.2
rsa==4.9.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
